Revert "#16 Rename to Practica"
This reverts commit 1c9b528923043058bc6b6246607e33fb6f95deff.
