+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.practica",
+    package = "com.guff",
 )
 
 android_resource(
     name = "res",
-    package = "com.practica",
+    package = "com.guff",
     res = "src/main/res",
 )
