+++ android/app/src/main/java/com/guff/MainActivity.java
@@ -1,4 +1,4 @@
-package com.practica;
+package com.guff;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "practica";
+        return "guff";
     }
 }
