+++ android/app/src/main/java/com/guff/MainApplication.java
@@ -1,4 +1,4 @@
-package com.practica;
+package com.guff;
 
 import android.app.Application;
