+++ android/settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = 'practica'
+rootProject.name = 'guff'
 
 include ':app'
