+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "practica",
-  "displayName": "practica"
+  "name": "guff",
+  "displayName": "guff"
 }
