+++ ios/guff.xcodeproj/xcshareddata/xcschemes/guff-tvOS.xcscheme
@@ -29,9 +29,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-               BuildableName = "practica-tvOS.app"
-               BlueprintName = "practica-tvOS"
-               ReferencedContainer = "container:practica.xcodeproj">
+               BuildableName = "guff-tvOS.app"
+               BlueprintName = "guff-tvOS"
+               ReferencedContainer = "container:guff.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
@@ -43,9 +43,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "practica-tvOSTests.xctest"
-               BlueprintName = "practica-tvOSTests"
-               ReferencedContainer = "container:practica.xcodeproj">
+               BuildableName = "guff-tvOSTests.xctest"
+               BlueprintName = "guff-tvOSTests"
+               ReferencedContainer = "container:guff.xcodeproj">
             </BuildableReference>
          </BuildActionEntry>
       </BuildActionEntries>
@@ -61,9 +61,9 @@
             <BuildableReference
                BuildableIdentifier = "primary"
                BlueprintIdentifier = "2D02E48F1E0B4A5D006451C7"
-               BuildableName = "practica-tvOSTests.xctest"
-               BlueprintName = "practica-tvOSTests"
-               ReferencedContainer = "container:practica.xcodeproj">
+               BuildableName = "guff-tvOSTests.xctest"
+               BlueprintName = "guff-tvOSTests"
+               ReferencedContainer = "container:guff.xcodeproj">
             </BuildableReference>
          </TestableReference>
       </Testables>
@@ -71,9 +71,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "practica-tvOS.app"
-            BlueprintName = "practica-tvOS"
-            ReferencedContainer = "container:practica.xcodeproj">
+            BuildableName = "guff-tvOS.app"
+            BlueprintName = "guff-tvOS"
+            ReferencedContainer = "container:guff.xcodeproj">
          </BuildableReference>
       </MacroExpansion>
       <AdditionalOptions>
@@ -94,9 +94,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "practica-tvOS.app"
-            BlueprintName = "practica-tvOS"
-            ReferencedContainer = "container:practica.xcodeproj">
+            BuildableName = "guff-tvOS.app"
+            BlueprintName = "guff-tvOS"
+            ReferencedContainer = "container:guff.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
       <AdditionalOptions>
@@ -113,9 +113,9 @@
          <BuildableReference
             BuildableIdentifier = "primary"
             BlueprintIdentifier = "2D02E47A1E0B4A5D006451C7"
-            BuildableName = "practica-tvOS.app"
-            BlueprintName = "practica-tvOS"
-            ReferencedContainer = "container:practica.xcodeproj">
+            BuildableName = "guff-tvOS.app"
+            BlueprintName = "guff-tvOS"
+            ReferencedContainer = "container:guff.xcodeproj">
          </BuildableReference>
       </BuildableProductRunnable>
    </ProfileAction>
