+++ ios/guffTests/guffTests.m
@@ -16,11 +16,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface practicaTests : XCTestCase
+@interface guffTests : XCTestCase
 
 @end
 
-@implementation practicaTests
+@implementation guffTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
