+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "practica",
+  "name": "guff",
   "version": "0.0.1",
   "private": true,
   "scripts": {
