+++ src/components/authentication/signup.js
@@ -24,7 +24,7 @@ module.exports = React.createClass({
     }
   },
   componentWillMount: function(){
-    AsyncStorage.removeItem('@practica:username');
+    AsyncStorage.removeItem('@guff:username');
   },
   render: function() {
     return (
@@ -71,7 +71,7 @@ module.exports = React.createClass({
           username: ''
         });
 
-        AsyncStorage.setItem('@practica:username', response.email);
+        AsyncStorage.setItem('@guff:username', response.email);
 
         this.props.navigator.immediatelyResetRouteStack([
           {name: 'tweets'}
